给数据的的操作细分
