+++ Gnote/Dialog/NewNotebook.resx
