--- Gnote/Model/MysqlBase.cs
@@ -1,34 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using MySql.Data;
-using MySql.Data.MySqlClient;
-
-namespace Gnote.Model
-{
-    class MysqlBase
-    {
-        public void base_mysql(string sql)
-        {
-            string MysqlConnectString = "Server=localhost;Database=gnote;Uid=root;Pwd=root";
-            MySqlConnection connection = new MySqlConnection(MysqlConnectString);
-            MySqlCommand cmd;
-            connection.Open();
-            try
-            {
-                cmd = connection.CreateCommand();
-                cmd.CommandText = sql;
-                cmd.ExecuteNonQuery();
-            }
-            catch (Exception)
-            {
-            }
-            finally
-            {
-                if (connection.State == System.Data.ConnectionState.Open)
-                    connection.Close();
-            }
-        }
-    }
-}
