--- Gnote/Open.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Gnote
-{
-    class Open
-    {
-    }
-}
